Use correct run-time path to git2cpp binary
