+++ test/test_git.py
@@ -3,16 +3,16 @@ import subprocess
 
 
 @pytest.mark.parametrize("arg", ['-v', '--version'])
-def test_version(arg):
-    cmd = ['build/git2cpp', arg]
+def test_version(git2cpp_path, arg):
+    cmd = [git2cpp_path, arg]
     p = subprocess.run(cmd, capture_output=True)
     assert p.returncode == 0
     assert p.stderr == b''
     assert p.stdout.startswith(b'git2cpp ')
 
 
-def test_error_on_unknown_option():
-    cmd = ['build/git2cpp', '--unknown']
+def test_error_on_unknown_option(git2cpp_path):
+    cmd = [git2cpp_path, '--unknown']
     p = subprocess.run(cmd, capture_output=True)
     assert p.returncode == 1
     assert p.stdout == b''
